pyproject.toml
[project]
name = "spongebob-machine"
version = "0.1.0"
description = "A Telegram bot that sends SpongeBob memes based on user emotions and intents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.1",
    "boto3>=1.42.28",
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "openpyxl>=3.1.5",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "sentence-transformers>=5.2.0",
    "sqlalchemy>=2.0.45",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bot", "src/nlp", "src/meme", "src/db"]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
